Refactor RoutingManager
